Extract formats to `beam.utils.json` package to be able to reuse them
